+++ app/src/main/java/woowacourse/shopping/data/model/response/AvailableTime.kt
@@ -1,0 +1,12 @@
+package woowacourse.shopping.data.model.response
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class AvailableTime(
+    @SerialName("start")
+    val start: String,
+    @SerialName("end")
+    val end: String,
+)
